+++ parser_test.go
@@ -12,11 +12,13 @@ import (
 	"github.com/dgrijalva/jwt-go/test"
 )
 
+var keyFuncError error = fmt.Errorf("error loading key")
+
 var (
 	jwtTestDefaultKey *rsa.PublicKey
 	defaultKeyFunc    jwt.Keyfunc = func(t *jwt.Token) (interface{}, error) { return jwtTestDefaultKey, nil }
 	emptyKeyFunc      jwt.Keyfunc = func(t *jwt.Token) (interface{}, error) { return nil, nil }
-	errorKeyFunc      jwt.Keyfunc = func(t *jwt.Token) (interface{}, error) { return nil, fmt.Errorf("error loading key") }
+	errorKeyFunc      jwt.Keyfunc = func(t *jwt.Token) (interface{}, error) { return nil, keyFuncError }
 	nilKeyFunc        jwt.Keyfunc = nil
 )
 
@@ -218,9 +220,15 @@ func TestParser_Parse(t *testing.T) {
 			if err == nil {
 				t.Errorf("[%v] Expecting error.  Didn't get one.", data.name)
 			} else {
+
+				ve := err.(*jwt.ValidationError)
 				// compare the bitfield part of the error
-				if e := err.(*jwt.ValidationError).Errors; e != data.errors {
+				if e := ve.Errors; e != data.errors {
 					t.Errorf("[%v] Errors don't match expectation.  %v != %v", data.name, e, data.errors)
+				}
+
+				if err.Error() == keyFuncError.Error() && ve.Inner != keyFuncError {
+					t.Errorf("[%v] Inner error does not match expectation.  %v != %v", data.name, ve.Inner, keyFuncError)
 				}
 			}
 		}