+++ VERSION_HISTORY.md
@@ -17,9 +17,15 @@
 	* Moved examples from README to executable example files
 	* Signing method registry is now thread safe
 
-#### 2.5.0
+#### 2.6.0
 
 This will likely be the last backwards compatible release before 3.0.0.
+
+* Exposed inner error within ValidationError
+* Fixed validation errors when using UseJSONNumber flag
+* Added several unit tests
+
+#### 2.5.0
 
 * Added support for signing method none.  You shouldn't use this.  The API tries to make this clear.
 * Updated/fixed some documentation